+++ src/fields/broadcastInlinePreview.ts
@@ -1,0 +1,13 @@
+import type { Field } from 'payload'
+
+export const createBroadcastInlinePreviewField = (): Field => {
+  return {
+    name: 'broadcastInlinePreview',
+    type: 'ui',
+    admin: {
+      components: {
+        Field: '/src/components/Broadcasts/BroadcastInlinePreview',
+      },
+    },
+  }
+}